pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "commi"
version = "0.0.1"
description = "Creates an AI-Powered commit messages"
authors = [
    "Timur Valiev <cptchunk@yandex.ru>"
]
license = "MIT"
readme = "README.md"

classifiers = [
  "Development Status :: 3 - Alpha",  # TODO
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Typing :: Typed",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]

homepage = "https://github.com/imtoopunkforyou/commi"

keywords = [
]  # TODO

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.24.2"
llama-cpp-python = "^0.3.20"
gitpython = "^3.1.47"

[tool.poetry.scripts]
commi = "commi.cli:main"

[tool.poetry.group.lint.dependencies]
ruff= "^0.15.12"
wemake-python-styleguide="^1.6.1"
flake8-digit-separator="^1.0.6"
codespell="^2.4.2"
mypy="^1.20.2"
pydantic = "^2.13.3"
pymarkdownlnt = "^0.9.36"

[tool.poetry.group.tests.dependencies]
pytest="^9.0.3"
pytest-sugar="^1.1.1"
